Galon/requirements.txt
wheel==0.37.1
pyrogram==1.4.16
tgcrypto==1.2.3
python-dotenv==0.20.0
googletrans==4.0.0rc1
aiohttp==3.8.1
aiofiles==0.8.0
redis==4.3.4
speedtest-cli==2.1.3
requests==2.28.1
beautifulsoup4==4.11.1
heroku3==5.1.4
pytgcalls==3.0.0.dev22
telegraph==2.1.0
pillow==9.2.0
youtube-search-python==1.5.1
yt-dlp==2022.7.18
wget==3.2
